+++ src/lambda_function.py
@@ -55,9 +55,9 @@ def lambda_handler(event, context):
 
     #extract send to slack channel from args
     sendto_data = None
-    if "sendto:" in text:
-        sendto_data = filter(None, text[text.index("sendto:") + 1:])
-        text = text[:text.index("sendto:")]
+    if "sendto" in text:
+        sendto_data = filter(None, text[text.index("sendto") + 1:])
+        text = text[:text.index("sendto")]
 
 
     if param_map['token'] != incoming_token:  # Check for a valid Slack token
@@ -111,6 +111,7 @@ def post_to_slack(val):
         r = requests.post(slack_response_url, json=payload)
 
 def send_to_slack(val, slack_channel):
+
     if isinstance(val, basestring):
         payload = {
         "text": query + "\n" + val,
@@ -118,6 +119,13 @@ def send_to_slack(val, slack_channel):
         }
         r = requests.post(slack_response_url, json=payload)
     else:
+        #slack parses out the # and @ chars into the below sequences,
+        # this reattaches the correct char to reform slack channel
+        if "%23" in slack_channel:
+            slack_channel.replace("%23", "#")
+        elif "%40" in slack_channel:
+            slack_channel.replace("%40", "@")
+
         payload = {
         "text": query,
         "attachments": val,
